requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
transformers
torch
gradio
scikit-learn
pickle-mixin 